Lab Completed
Thinking problem completed, the program runs without errors, and the tests return the expected values.

I will keep the unit tests in main and make a branch for turning in the project without unit tests, because I fear the mighty grade hammer, and the penalty for doing things beyond what is expected.